requirements.txt
click>=8.0
python-magic>=0.4.27
rich>=13.0
